[refactor] added a util file
This has the subroutines.
