+++ Makefile.am
@@ -1,6 +1,6 @@
 
-ACLOCAL_AMFLAGS = -I m4 
-AM_CPPFLAGS = -I$(top_srcdir)/include -std=c++11 
+ACLOCAL_AMFLAGS = -I m4
+AM_CPPFLAGS = -I$(top_srcdir)/include -std=c++11
 
 instfilterdir = $(datarootdir)/castus4/filterscripts
 pkgconfiglibdir = $(libdir)/pkgconfig
@@ -47,13 +47,13 @@ libcastus4_public_la_SOURCES = \
     src/lib/schedule_helpers.cpp \
     src/lib/c_schedule.cpp
 
-filter_example_check_duration_SOURCES = filter-example-check-duration.cpp
+filter_example_check_duration_SOURCES = filter-example-check-duration.cpp utils.cpp
 filter_example_check_duration_LDADD = libcastus4-public.la
 
-filter_example_check_duration_and_ripple_SOURCES = filter-example-check-duration-and-ripple.cpp
+filter_example_check_duration_and_ripple_SOURCES = filter-example-check-duration-and-ripple.cpp utils.cpp
 filter_example_check_duration_and_ripple_LDADD = libcastus4-public.la
 
-filter_example_check_duration_and_ripple_bidi_SOURCES = filter-example-check-duration-and-ripple-bidi.cpp
+filter_example_check_duration_and_ripple_bidi_SOURCES = filter-example-check-duration-and-ripple-bidi.cpp utils.cpp
 filter_example_check_duration_and_ripple_bidi_LDADD = libcastus4-public.la
 
 castus4_public_demo_parsetime_SOURCES = src/bin/parsetime.cpp
