+++ include/castus4-public/chomp.h
@@ -1,4 +1,7 @@
 
+#ifndef Castus4publicChomp_h
+#define Castus4publicChomp_h
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -9,4 +12,4 @@ void castus4public_chomp(char *s);
 }
 #endif
 
-
+#endif
