+++ include/castus4-public/gentime.h
@@ -1,4 +1,7 @@
 
+#ifndef Castus4publicGentime_h
+#define Castus4publicGentime_h
+
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <assert.h>
@@ -13,3 +16,4 @@
 
 std::string castus4_schedule_print_time(int stm_type,const struct tm *stm,unsigned long stm_us=0);
 
+#endif
