+++ include/castus4-public/schedule.h
@@ -1,4 +1,7 @@
 
+#ifndef Castus4publicSchedule_h
+#define Castus4publicSchedule_h
+
 enum {
 	C4_SCHED_TYPE_NONE=-1,
 	C4_SCHED_TYPE_DAILY=0,
@@ -8,3 +11,4 @@ enum {
 	C4_SCHED_TYPE_INTERVAL=4
 };
 
+#endif
